+++ packages.h
@@ -1,12 +1,14 @@
 #pragma once
 
+#include <stdint.h>
+#include <stdio.h>
+
 // Is it worth wrapping this in a struct?
 // Yeah, probably just to future-proof it.
 typedef struct {
-  unsigned int to_update;
+  uint_fast16_t to_update;
 } PackagesInfo;
 
 void update_packages_info(PackagesInfo *, FILE *);
 
 PackagesInfo init_packages_info(void);
-
